// apps/web/src/api/generated/models/invitationCreate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * RSP API
 * Same-origin API for the Ravi's Study Program application.
 * OpenAPI spec version: 2.0.0
 */
import type { InvitationCreateRole } from './invitationCreateRole';

export interface InvitationCreate {
  /**
   * @minLength 1
   * @maxLength 100
   */
  name: string;
  email: string;
  role: InvitationCreateRole;
}